+++ app/services/import/fetchers/github/pull_requests.rb
@@ -6,6 +6,8 @@ module Import
       class PullRequests
         prepend ApplicationService
 
+        PER_PAGE = 100
+
         def initialize(repository:, fetch_client: GithubApi::Client, duration_days: Insight::FETCH_DAYS_PERIOD)
           @fetch_client = fetch_client.new(repository: repository)
           @started_at_limit = (DateTime.now - (duration_days.days * (repository.premium? ? 2 : 1))).beginning_of_day
@@ -18,11 +20,13 @@ module Import
           loop do
             # default sorting is desc by created_at attribute
             # first comes newest PRs
-            result = @fetch_client.pull_requests(params: { state: 'all', per_page: 100, page: page })
+            result = @fetch_client.pull_requests(params: { state: 'all', per_page: PER_PAGE, page: page })
             result = filter_result(result)
             break if result.blank?
 
             @result.concat(result)
+            break if result.size != PER_PAGE
+
             page += 1
           end
         end