+++ db/migrate/20230125135040_remove_open_attribute_from_pull_requests.rb
@@ -1,0 +1,9 @@
+class RemoveOpenAttributeFromPullRequests < ActiveRecord::Migration[7.0]
+  def up
+    safety_assured { remove_column :pull_requests, :open }
+  end
+
+  def down
+    add_column :pull_requests, :open, :boolean, null: false, default: true
+  end
+end