+++ db/structure.sql
@@ -459,8 +459,7 @@ CREATE TABLE public.pull_requests (
     pull_number integer NOT NULL,
     pull_created_at timestamp(6) without time zone NOT NULL,
     pull_closed_at timestamp(6) without time zone,
-    pull_merged_at timestamp(6) without time zone,
-    open boolean DEFAULT true NOT NULL
+    pull_merged_at timestamp(6) without time zone
 );
 
 
@@ -1276,6 +1275,7 @@ INSERT INTO "schema_migrations" (version) VALUES
 ('20230120143333'),
 ('20230120194630'),
 ('20230121154845'),
-('20230121185501');
+('20230121185501'),
+('20230125135040');