+++ spec/services/import/fetchers/github/reviews_spec.rb
@@ -45,10 +45,7 @@ describe Import::Fetchers::Github::Reviews, type: :service do
   before do
     allow(fetch_client).to receive(:new).and_return(fetch_service)
     allow(fetch_service).to(
-      receive(:pull_request_reviews).with(pull_number: 1, params: { per_page: 25, page: 1 }).and_return(data)
-    )
-    allow(fetch_service).to(
-      receive(:pull_request_reviews).with(pull_number: 1, params: { per_page: 25, page: 2 }).and_return([])
+      receive(:pull_request_reviews).with(pull_number: 1, params: { per_page: 50, page: 1 }).and_return(data)
     )
   end